Delay VkScript Shader Compilation.
This CL changes VkScript to delay shader compilation until execution,
instead of at parse time. This allows VkScript to share the Script
object with AmberScript.
