+++ src/CMakeLists.txt
@@ -19,7 +19,6 @@ set(AMBER_SOURCES
     amberscript/parser.cc
     amberscript/pipeline.cc
     amberscript/script.cc
-    amberscript/shader.cc
     buffer.cc
     command.cc
     command_data.cc
@@ -31,6 +30,7 @@ set(AMBER_SOURCES
     pipeline_data.cc
     result.cc
     script.cc
+    shader.cc
     shader_compiler.cc
     tokenizer.cc
     value.cc
