+++ src/amberscript/script_test.cc
@@ -17,9 +17,9 @@
 #include "gtest/gtest.h"
 #include "src/amberscript/pipeline.h"
 #include "src/amberscript/script.h"
-#include "src/amberscript/shader.h"
 #include "src/buffer.h"
 #include "src/make_unique.h"
+#include "src/shader.h"
 
 namespace amber {
 namespace amberscript {
